# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Treffin API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /articles:
    get:
      operationId: getArticles
      tags: [articles]
      summary: Get list of articles
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [trending, newest, most_liked]
      responses:
        "200":
          description: Articles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create a new article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}/annotations:
    get:
      operationId: getArticleAnnotations
      tags: [articles]
      summary: Get annotations for an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of annotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Annotation"
    post:
      operationId: createAnnotation
      tags: [articles]
      summary: Create an annotation on an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationInput"
      responses:
        "201":
          description: Created annotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
        "401":
          description: Unauthorized
  /annotations/{id}:
    delete:
      operationId: deleteAnnotation
      tags: [articles]
      summary: Delete an annotation (owner or article author)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /articles/{id}/like:
    post:
      operationId: likeArticle
      tags: [articles]
      summary: Like an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Article not found
  /articles/{id}/comments:
    post:
      operationId: createArticleComment
      tags: [articles]
      summary: Post a comment on an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebateCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebateComment"
        "404":
          description: Article not found
...
  /articles/{id}/review-request:
    post:
      operationId: submitReviewRequest
      tags: [articles]
      summary: Submit an article for peer review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Review request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleReviewRequest"
        "409":
          description: Review request already exists
...
components:
  schemas:
...
    Article:
      type: object
      required: [id, title, authorId, authorName, authorTitle, authorAvatar, createdAt, readTime, likes, isTrending, isFeatured, isExpertReviewed]
      properties:
        id:
          type: integer
        title:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        authorId:
          type: integer
        authorName:
          type: string
        authorTitle:
          type: string
        authorAvatar:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        readTime:
          type: integer
        likes:
          type: integer
        liked:
          type: boolean
        isVerified:
          type: boolean
        createdAt:
          type: string
        isTrending:
          type: boolean
        isFeatured:
          type: boolean
        isExpertReviewed:
          type: boolean
        reviewRequestStatus:
          type: ["string", "null"]
        editedAt:
          type: ["string", "null"]
        isRemoved:
          type: boolean
        status:
          type: string
    ArticleInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        peerReview:
          type: boolean
    Annotation:
      type: object
      required: [id, articleId, userId, selectedText, comment, paragraphIndex, createdAt, authorName, authorAvatar]
      properties:
        id:
          type: integer
        articleId:
          type: integer
        userId:
          type: integer
        selectedText:
          type: string
        comment:
          type: string
        paragraphIndex:
          type: integer
        createdAt:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
    AnnotationInput:
      type: object
      required: [selectedText, comment, paragraphIndex]
      properties:
        selectedText:
          type: string
        comment:
          type: string
        paragraphIndex:
          type: integer
...
    DebateComment:
      type: object
      required: [id, authorId, authorName, content, createdAt, isRemoved]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        side:
          type: ["string", "null"]
        isFlagged:
          type: boolean
        flagLabel:
          type: ["string", "null"]
        createdAt:
          type: string
        editedAt:
          type: ["string", "null"]
        isRemoved:
          type: boolean
        removedReason:
          type: ["string", "null"]
        sources:
          type: ["string", "null"]
          description: JSON array of {url, label} source citations
        wordCount:
          type: ["integer", "null"]
        personalAttackWarning:
          type: ["string", "null"]
          description: Non-blocking advisory returned when post contains second-person attack patterns
        likes:
          type: integer
          description: Total like count for this comment
        likedByMe:
          type: boolean
          description: Whether the authenticated user has liked this comment
        parentCommentId:
          type: ["integer", "null"]
          description: ID of the parent comment if this is a reply
...
    ArticleReviewRequest:
      type: object
      required: [id, articleId, requesterId, status, createdAt]
      properties:
        id:
          type: integer
        articleId:
          type: integer
        requesterId:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewerNote:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DebateCommentInput:
      type: object
      required: [authorName, content]
      properties:
        authorId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        side:
          type: string
        sources:
          type: string
          description: JSON array of {url, label} source citations
        parentCommentId:
          type: integer
          description: ID of parent comment when posting a reply
...